Most of the functionality for guest users work except for checkout.

+++ browser/js/cart/cart.controller.js
@@ -1,5 +1,11 @@
 app.controller('CartCtrl', function ($scope, $state, cart, CartService, OrderService, ProductService) {
-    $scope.cart = cart;
+    console.log("Cart in cart controller ", cart);
+    if(localStorage.getItem('cart')){
+      $scope.cart = JSON.parse(localStorage.getItem('cart'));
+      console.log($scope.cart);
+    }else{
+      $scope.cart = cart;
+    }
     $scope.uniqueProducts = [];
     $scope.editing = false;
     $scope.total = 0;
@@ -10,8 +16,16 @@ app.controller('CartCtrl', function ($scope, $state, cart, CartService, OrderSer
     };
     $scope.emptyCart = function () {
         $scope.uniqueProducts = [];
-        $scope.updateCart();
         $scope.total = 0;
+        if(localStorage.getItem('cart')){
+          console.log("empty local Storage cart");
+          console.dir("localStorage Cart ", JSON.parse(localStorage.getItem('cart')))
+          $scope.cart.contents = [];
+          localStorage.setItem('cart',JSON.stringify($scope.cart));
+          CartService.getLocalCart();
+        }else{
+          $scope.updateCart();
+        }
     };
     $scope.showCheckOut = function () {
         $scope.checkOutView = !$scope.checkOutView;
@@ -26,10 +40,15 @@ app.controller('CartCtrl', function ($scope, $state, cart, CartService, OrderSer
         };
         // I think we have to .then off of this since it is async
         $scope.updateProducts();
-        OrderService.createOrder(user,order).then(function(savedOrder){
-            $scope.emptyCart();
-            $scope.checkOutView = false;
-        });
+        if($scope.cart.localStorage){
+          console.log("Checkout on localStorage");
+        }else{
+          console.log("Checkout on user");
+          OrderService.createOrder(user,order).then(function(savedOrder){
+              $scope.emptyCart();
+              $scope.checkOutView = false;
+          });
+        }
     };
     $scope.updateProducts = function(){
         $scope.uniqueProducts.forEach(function(product){
@@ -55,16 +74,31 @@ app.controller('CartCtrl', function ($scope, $state, cart, CartService, OrderSer
         });
     };
     $scope.removeFromCart = function (product) {
-        CartService.removeFromCart(product, $scope.cart._id)
-            .then(function (updatedCart) {
-                $scope.cart = updatedCart;
-                $state.reload();
-            });
+        if(localStorage.getItem('cart')){
+          var newContents = [];
+          var cart = JSON.parse(localStorage.getItem('cart'));
+          cart.contents.forEach(
+            function(currentProduct){
+              if(currentProduct._id !== product._id) newContents.push(currentProduct);
+            }
+          );
+          cart.contents = newContents;
+          localStorage.setItem("cart", JSON.stringify(cart));
+          CartService.getLocalCart();
+          $state.reload();
+        }else{
+          CartService.removeFromCart(product, $scope.cart._id)
+          .then(
+            function (updatedCart) {
+              $scope.cart = updatedCart;
+              $state.reload();
+            }
+          );
+        }
     };
     $scope.increaseQuantity = function (product) {
         $scope.editing = true;
         product.cartQuantity++;
-        
     };
     $scope.reduceQuantity = function (product) {
         $scope.editing = true;
